backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.32
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-dotenv==0.19.2
alembic==1.7.7